requirements.txt
aiohttp>=3.8.0
cloudflare>=2.19.2
python-dotenv>=1.0.0
tqdm>=4.64.0
requests>=2.31.0
